+++ libs/blkstructs/src/testing/utils.rs
@@ -7,7 +7,7 @@ pub fn random_valid_txx(
     start_coindata: CoinData,
     signer: tmelcrypt::Ed25519SK,
     cons: &melscript::Script,
-    fee: u64
+    fee: u128
 ) -> Vec<Transaction> {
     let mut pqueue: BinaryHeap<(u64, CoinID, CoinData)> = BinaryHeap::new();
     pqueue.push((rng.gen(), start_coin, start_coindata));
